Move code from JavaToDafnyCompiler, to ClassCompiler
