+++ verifier/src/main/java/com/aws/jverify/verifier/BlockCompiler.java
@@ -349,7 +349,7 @@ public class BlockCompiler {
 
             var baseConstructorName = compiler.nameCompiler.getCompiledName(baseConstructor);
             var baseConstructorClassName = compiler.nameCompiler.getCompiledName(baseConstructor.enclClass());
-            var initName = compiler.getInitMethodName(baseConstructorClassName, baseConstructorName);
+            var initName = compiler.nameCompiler.getInitMethodName(baseConstructorClassName, baseConstructorName);
             var arguments = invocation.getArguments().stream().map(
                     e -> new ActualBinding(null, compiler.expressionCompiler.toExpr(e), false)).toList();
             var applySuffix = new ApplySuffix(origin,
